refactor : #90 PreAuthorize 어노테이션 적용 - PreAuthorize 어노테이션 적용 - BucketConfig 위치 이동
