+++ src/main/java/com/sparta/delivery_app/domain/menu/controller/MenuController.java
@@ -12,6 +12,7 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -29,6 +30,7 @@ public class MenuController {
      * @param user
      * @return 성공 메시지 및 등록된 메뉴 정보
      */
+    @PreAuthorize("hasRole('MANAGER')")
     @PostMapping
     public ResponseEntity<RestApiResponse<MenuAddResponseDto>> menuAdd(
             @Valid @RequestBody final MenuAddRequestDto requestDto,
@@ -47,6 +49,7 @@ public class MenuController {
      * @param user
      * @return 성공 메시지 및 수정된 메뉴 정보
      */
+    @PreAuthorize("hasRole('MANAGER')")
     @PutMapping("/{menuId}")
     public ResponseEntity<RestApiResponse<MenuModifyResponseDto>> menuModify(
             @PathVariable Long menuId,
@@ -64,6 +67,7 @@ public class MenuController {
      * @param menuId
      * @param user
      */
+    @PreAuthorize("hasRole('MANAGER')")
     @DeleteMapping("/{menuId}")
     public ResponseEntity<RestApiResponse<Void>> menuDelete(
             @PathVariable Long menuId,
