+++ src/main/java/com/sparta/delivery_app/domain/menu/service/MenuService.java
@@ -35,7 +35,7 @@ public class MenuService {
      * @return responseDto
      */
     public MenuAddResponseDto addMenu(final MenuAddRequestDto requestDto, AuthenticationUser user) {
-        User findUser = checkUserAuth(user);
+        User findUser = userAdaptor.queryUserByEmailAndStatus(user.getUsername());
 
         Store findStore = storeAdaptor.queryStoreId(findUser);
 
@@ -54,7 +54,7 @@ public class MenuService {
      */
     @Transactional
     public MenuModifyResponseDto modifyMenu(Long menuId, final MenuModifyRequestDto requestDto, AuthenticationUser user) {
-        User findUser = checkUserAuth(user);
+        User findUser = userAdaptor.queryUserByEmailAndStatus(user.getUsername());
 
         Store store = storeAdaptor.queryStoreId(findUser);
         Menu menu = menuAdaptor.queryMenuByIdAndMenuStatus(menuId);
@@ -72,7 +72,7 @@ public class MenuService {
      */
     @Transactional
     public void deleteMenu(Long menuId, AuthenticationUser user) {
-        User findUser = checkUserAuth(user);
+        User findUser = userAdaptor.queryUserByEmailAndStatus(user.getUsername());
 
         Store store = storeAdaptor.queryStoreId(findUser);
         Menu menu = menuAdaptor.queryMenuByIdAndMenuStatus(menuId);
@@ -90,17 +90,5 @@ public class MenuService {
         if(!menu.getStore().getId().equals(storeId)) {
             throw new StoreMenuMismatchException(OrderErrorCode.STORE_MENU_MISMATCH);
         }
-    }
-
-    /**
-     * 사용자 인증
-     * @param user
-     * @return
-     */
-    private User checkUserAuth(AuthenticationUser user) {
-        User findUser = userAdaptor.queryUserByEmailAndStatus(user.getUsername());
-        checkManagerEnable(findUser);
-
-        return findUser;
     }
 }
